Latest improvements

+++ Projects/todo_list/todo_list_backend/src/main/java/com/todoList/dao/image/ImageDAOImpl.java
@@ -1,5 +1,6 @@
 package com.todoList.dao.image;
 
+import com.todoList.AOP.Exceptions.ExceptionObjects.NotFoundException;
 import com.todoList.entities.Image;
 import jakarta.persistence.EntityManager;
 import lombok.AllArgsConstructor;
@@ -12,7 +13,7 @@ public class ImageDAOImpl implements ImageDAO {
     private EntityManager entityManager;
 
     @Override
-    public Image get(long id) {
+    public Image get(long id){
         return entityManager.find(Image.class, id);
     }
 
@@ -22,10 +23,12 @@ public class ImageDAOImpl implements ImageDAO {
     }
 
     @Override
-    public void delete(long id) {
+    public void delete(long id) throws NotFoundException {
         Image image = get(id);
-        image.setUser(null);
-        entityManager.remove(image);
+        if(image != null) {
+            image.setUser(null);
+            entityManager.remove(image);
+        }
     }
 
 }
